moved test spec to testcases, moved components to utils, and updated README.md with frontend test skel
fixed imports for types and components

fixed indentation and updated README.md with more specific frontend test skeletons and implemented examples

fixed typo in TOC
