+++ cypress/utils/components/labeled-select.po.ts
@@ -1,4 +1,4 @@
-import ComponentPo from '@/pageobjects/components/component.po';
+import ComponentPo from './component.po';
 
 export default class LabeledSelectPo extends ComponentPo {
   select(option: string | undefined) {
